Calculate the censored geometric mean

Calculate the censored geometric mean

Usage

geomean_tobit(x, lower = 0, upper = Inf, na.rm = TRUE, ...)

Arguments

x

A numeric vector

lower, upper

The lower and upper bounds for censoring x

na.rm

A logical value indicating whether NA values should be stripped from x before the computation proceeds.

...

Arguments passed on to mean_tobit

lower,upper

The lower and upper bounds for censoring x

include_equal,include_equal_lower,include_equal_upper

Should the bounds consider x <= lower and x >= upper (if TRUE) or x < lower and x > upper (if FALSE) to be censored? (include_equal is overridden by setting include_equal_lower or include_equal_upper.)

distribution

What distribution (Student "t" or "normal") should be used. If "t", the degrees of freedom are the number of non-NA values in x.

Value

A "sd_tobit" class object

See also